Add entitlements system
