+++ packages/web/src/ee/features/searchContexts/syncSearchContexts.ts
@@ -1,13 +1,21 @@
 import { env } from "@/env.mjs";
+import { getPlan, hasEntitlement } from "@/features/entitlements/server";
 import { SINGLE_TENANT_ORG_ID } from "@/lib/constants";
 import { prisma } from "@/prisma";
 import { SearchContext } from "@sourcebot/schemas/v3/index.type";
 import micromatch from "micromatch";
 
-
 export const syncSearchContexts = async (contexts?: { [key: string]: SearchContext }) => {
     if (env.SOURCEBOT_TENANCY_MODE !== 'single') {
-        throw new Error("Search contexts are not supported in this tenancy mode");
+        throw new Error("Search contexts are not supported in this tenancy mode. Set SOURCEBOT_TENANCY_MODE=single in your environment variables.");
+    }
+
+    if (!hasEntitlement("search-contexts")) {
+        if (contexts) {
+            const plan = getPlan();
+            console.error(`Search contexts are not supported in your current plan: ${plan}. If you have a valid enterprise license key, pass it via SOURCEBOT_EE_LICENSE_KEY. For support, contact team@sourcebot.dev.`);
+        }
+        return;
     }
 
     if (contexts) {
