+++ packages/web/src/features/entitlements/README.md
@@ -1,0 +1,8 @@
+# Entitlements
+
+Entitlements control the availability of certain features dependent on the current plan. Entitlements are managed at the **instance** level.
+
+Some definitions:
+
+- `Plan`: A plan is a tier of features. Examples: `oss`, `cloud:team`, `self-hosted:enterprise`.
+- `Entitlement`: An entitlement is a feature that is available to a instance. Examples: `search-contexts`, `billing`.
