+++ packages/web/src/features/entitlements/constants.ts
@@ -2,7 +2,6 @@
 const planLabels = {
     oss: "OSS",
     "cloud:team": "Team",
-    "cloud:enterprise": "Enterprise",
     "self-hosted:enterprise": "Enterprise (Self-Hosted)",
 } as const;
 export type Plan = keyof typeof planLabels;
@@ -16,6 +15,5 @@ export type Entitlement = (typeof entitlements)[number];
 export const entitlementsByPlan: Record<Plan, Entitlement[]> = {
     oss: [],
     "cloud:team": [],
-    "cloud:enterprise": ["search-contexts"],
     "self-hosted:enterprise": ["search-contexts"],
 } as const;
