+++ packages/web/src/features/entitlements/usePlan.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+import { PlanContext } from "./planProvider";
+
+export const usePlan = () => {
+    const plan = useContext(PlanContext);
+    return plan;
+}
